+++ tests/testthat/test-RushWorker.R
@@ -38,7 +38,7 @@ test_that("a worker is registered", {
   expect_data_table(worker_info, nrows = 1)
   expect_names(
     names(worker_info),
-    permutation.of = c("worker_id", "pid", "hostname", "profile", "heartbeat", "state")
+    permutation.of = c("worker_id", "pid", "hostname", "profile", "heartbeat", "restarted_from", "state")
   )
   expect_equal(worker_info$worker_id, rush$worker_id)
   expect_equal(worker_info$pid, Sys.getpid())